Move unittest into SYCL2020 subdir
