+++ sycl/unittests/SYCL2020/KernelBundle.cpp
@@ -7,6 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #include <CL/sycl.hpp>
+#include <detail/kernel_bundle_impl.hpp>
 
 #include <helpers/CommonRedefinitions.hpp>
 #include <helpers/PiImage.hpp>
@@ -89,4 +90,43 @@ TEST(KernelBundle, GetKernelBundleFromKernel) {
       Kernel.get_kernel_bundle();
 
   EXPECT_EQ(KernelBundle, RetKernelBundle);
+}
+
+TEST(KernelBundle, KernelBundleAndItsDevImageStateConsistency) {
+  sycl::platform Plt{sycl::default_selector()};
+  if (Plt.is_host()) {
+    std::cerr << "Test is not supported on host, skipping\n";
+    return; // test is not supported on host.
+  }
+
+  if (Plt.get_backend() == sycl::backend::cuda) {
+    std::cerr << "Test is not supported on CUDA platform, skipping\n";
+    return;
+  }
+
+  sycl::unittest::PiMock Mock{Plt};
+  setupDefaultMockAPIs(Mock);
+
+  const sycl::device Dev = Plt.get_devices()[0];
+
+  sycl::queue Queue{Dev};
+
+  const sycl::context Ctx = Queue.get_context();
+
+  sycl::kernel_bundle KernelBundle =
+      sycl::get_kernel_bundle<TestKernel, sycl::bundle_state::input>(Ctx, {Dev});
+
+  auto ObjBundle = sycl::compile(KernelBundle, KernelBundle.get_devices());
+  EXPECT_FALSE(ObjBundle.empty()) << "Expect non-empty obj kernel bundle";
+
+  auto ObjBundleImpl = sycl::detail::getSyclObjImpl(ObjBundle);
+  EXPECT_EQ(ObjBundleImpl->get_bundle_state(), sycl::bundle_state::object)
+      << "Expect object device image in bundle";
+
+  auto LinkBundle = sycl::link(ObjBundle, ObjBundle.get_devices());
+  EXPECT_FALSE(LinkBundle.empty()) << "Expect non-empty exec kernel bundle";
+
+  auto LinkBundleImpl = sycl::detail::getSyclObjImpl(LinkBundle);
+  EXPECT_EQ(LinkBundleImpl->get_bundle_state(), sycl::bundle_state::executable)
+      << "Expect executable device image in bundle";
 }
