--- sycl/unittests/misc/KernelBundle.cpp
@@ -1,123 +1,0 @@
-//===------- KernelBundle.cpp - Kernel bundle processing unit test --------===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===---------------------------------------------------------------------===//
-
-#define SYCL2020_DISABLE_DEPRECATION_WARNINGS
-
-#include <CL/sycl.hpp>
-#include <detail/kernel_bundle_impl.hpp>
-#include <gtest/gtest.h>
-#include <helpers/PiImage.hpp>
-#include <helpers/PiMock.hpp>
-
-class KBTKernel;
-static constexpr const char *KBTKernelName = "KBTKernel";
-
-// static std::string BuildOpts;
-__SYCL_INLINE_NAMESPACE(cl) {
-namespace sycl {
-namespace detail {
-template <> struct KernelInfo<KBTKernel> {
-  static constexpr unsigned getNumParams() { return 0; }
-  static const kernel_param_desc_t &getParamDesc(int) {
-    static kernel_param_desc_t Dummy;
-    return Dummy;
-  }
-  static constexpr const char *getName() { return KBTKernelName; }
-  static constexpr bool isESIMD() { return false; }
-  static constexpr bool callsThisItem() { return false; }
-  static constexpr bool callsAnyThisFreeFunction() { return false; }
-};
-
-} // namespace detail
-} // namespace sycl
-} // __SYCL_INLINE_NAMESPACE(cl)
-
-static pi_result redefinedProgramCreate(pi_context, const void *, size_t,
-                                        pi_program *) {
-  return PI_SUCCESS;
-}
-
-static pi_result redefinedProgramCompile(pi_program, pi_uint32,
-                                         const pi_device *, const char *options,
-                                         pi_uint32, const pi_program *,
-                                         const char **,
-                                         void (*)(pi_program, void *), void *) {
-  return PI_SUCCESS;
-}
-
-static pi_result redefinedProgramLink(pi_context, pi_uint32, const pi_device *,
-                                      const char *options, pi_uint32,
-                                      const pi_program *,
-                                      void (*)(pi_program, void *), void *,
-                                      pi_program *) {
-  return PI_SUCCESS;
-}
-
-static void setupDefaultMockAPIs(sycl::unittest::PiMock &Mock) {
-  using namespace sycl::detail;
-  Mock.redefine<PiApiKind::piProgramCreate>(redefinedProgramCreate);
-  Mock.redefine<PiApiKind::piProgramCompile>(redefinedProgramCompile);
-  Mock.redefine<PiApiKind::piProgramLink>(redefinedProgramLink);
-}
-
-static sycl::unittest::PiImage generateDefaultImage() {
-  using namespace sycl::unittest;
-
-  std::vector<unsigned char> Bin{0, 1, 2, 3, 4, 5}; // Random data
-  PiArray<PiOffloadEntry> Entries = makeEmptyKernels({KBTKernelName});
-
-  return PiImage{PI_DEVICE_BINARY_TYPE_SPIRV,            // Format
-                 __SYCL_PI_DEVICE_BINARY_TARGET_SPIRV64, // DeviceTargetSpec
-                 "",                                     // Compile options
-                 "",                                     // Link options
-                 std::move(Bin),
-                 std::move(Entries),
-                 {}}; // Property set
-}
-
-sycl::unittest::PiImage KBTImg = generateDefaultImage();
-sycl::unittest::PiImageArray<1> KBTImgArray{&KBTImg};
-
-TEST(KernelBundle, KernelBundleLink) {
-  sycl::platform Plt{sycl::default_selector()};
-  if (Plt.is_host()) {
-    std::cerr << "Test is not supported on host, skipping\n";
-    return; // test is not supported on host.
-  }
-
-  if (Plt.get_backend() == sycl::backend::cuda) {
-    std::cerr << "Test is not supported on CUDA platform, skipping\n";
-    return;
-  }
-
-  sycl::unittest::PiMock Mock{Plt};
-  setupDefaultMockAPIs(Mock);
-
-  const sycl::device Dev = Plt.get_devices()[0];
-
-  sycl::queue Queue{Dev};
-
-  const sycl::context Ctx = Queue.get_context();
-
-  sycl::kernel_bundle KernelBundle =
-      sycl::get_kernel_bundle<KBTKernel, sycl::bundle_state::input>(Ctx, {Dev});
-
-  auto ObjBundle = sycl::compile(KernelBundle, KernelBundle.get_devices());
-  EXPECT_FALSE(ObjBundle.empty()) << "Expect non-empty obj kernel bundle";
-
-  auto ObjBundleImpl = sycl::detail::getSyclObjImpl(ObjBundle);
-  EXPECT_EQ(ObjBundleImpl->get_bundle_state(), sycl::bundle_state::object)
-      << "Expect object device image in bundle";
-
-  auto LinkBundle = sycl::link(ObjBundle, ObjBundle.get_devices());
-  EXPECT_FALSE(LinkBundle.empty()) << "Expect non-empty exec kernel bundle";
-
-  auto LinkBundleImpl = sycl::detail::getSyclObjImpl(LinkBundle);
-  EXPECT_EQ(LinkBundleImpl->get_bundle_state(), sycl::bundle_state::executable)
-      << "Expect executable device image in bundle";
-}
